refact: move create repository and add roles out of developer_tool
