+++ taf/api/repository.py
@@ -1,0 +1,189 @@
+from functools import partial
+import click
+
+from collections import defaultdict
+from pathlib import Path
+from taf.api.metadata import update_target_metadata
+from taf.api.roles import _create_delegations, _initialize_roles_and_keystore, _role_obj
+
+from taf.auth_repo import AuthenticationRepository
+from taf.constants import DEFAULT_RSA_SIGNATURE_SCHEME, YUBIKEY_EXPIRATION_DATE
+from taf.exceptions import TargetsMetadataUpdateError
+from taf.git import GitRepository
+from taf.keys import get_key_name, load_sorted_keys_of_new_roles
+from taf.repository_tool import Repository, yubikey_signature_provider
+from tuf.repository_tool import create_new_repository
+
+
+def create_repository(
+    repo_path, keystore=None, roles_key_infos=None, commit=False, test=False
+):
+    """
+    <Purpose>
+        Create a new authentication repository. Generate initial metadata files.
+        The initial targets metadata file is empty (does not specify any targets).
+    <Arguments>
+        repo_path:
+        Authentication repository's location
+        targets_directory:
+        Directory which contains target repositories
+        keystore:
+        Location of the keystore files
+        roles_key_infos:
+        A dictionary whose keys are role names, while values contain information about the keys.
+        commit:
+        Indicates if the changes should be automatically committed
+        test:
+        Indicates if the created repository is a test authentication repository
+    """
+    yubikeys = defaultdict(dict)
+    auth_repo = AuthenticationRepository(path=repo_path)
+    repo_path = Path(repo_path)
+
+    if not _check_if_can_create_repository(auth_repo):
+        return
+
+    roles_key_infos, keystore = _initialize_roles_and_keystore(
+        roles_key_infos, keystore
+    )
+
+    repository = create_new_repository(str(auth_repo.path))
+    roles_infos = roles_key_infos.get("roles")
+    signing_keys, verification_keys = load_sorted_keys_of_new_roles(
+        auth_repo, roles_infos, repository, keystore, yubikeys
+    )
+    # set threshold and register keys of main roles
+    # we cannot do the same for the delegated roles until delegations are created
+    for role_name, role_key_info in roles_infos.items():
+        threshold = role_key_info.get("threshold", 1)
+        is_yubikey = role_key_info.get("yubikey", False)
+        _setup_role(
+            role_name,
+            threshold,
+            is_yubikey,
+            repository,
+            verification_keys[role_name],
+            signing_keys.get(role_name),
+        )
+
+    _create_delegations(roles_infos, repository, verification_keys, signing_keys)
+
+    # if the repository is a test repository, add a target file called test-auth-repo
+    if test:
+        test_auth_file = (
+            Path(auth_repo.path, auth_repo.targets_path) / auth_repo.TEST_REPO_FLAG_FILE
+        )
+        test_auth_file.touch()
+
+    # register and sign target files (if any)
+    try:
+        taf_repository = Repository(repo_path)
+        taf_repository._tuf_repository = repository
+        register_target_files(
+            repo_path, keystore, roles_key_infos, commit=commit, taf_repo=taf_repository
+        )
+    except TargetsMetadataUpdateError:
+        # if there are no target files
+        repository.writeall()
+
+    print("Created new authentication repository")
+
+    if commit:
+        auth_repo.init_repo()
+        commit_message = input("\nEnter commit message and press ENTER\n\n")
+        auth_repo.commit(commit_message)
+
+def _check_if_can_create_repository(auth_repo):
+    repo_path = Path(auth_repo.path)
+    if repo_path.is_dir():
+        # check if there is non-empty metadata directory
+        if auth_repo.metadata_path.is_dir() and any(auth_repo.metadata_path.iterdir()):
+            if auth_repo.is_git_repository:
+                print(
+                    f'"{repo_path}" is a git repository containing the metadata directory. Generating new metadata files could make the repository invalid. Aborting.'
+                )
+                return False
+            if not click.confirm(
+                f'Metadata directory found inside "{repo_path}". Recreate metadata files?'
+            ):
+                return False
+    return True
+
+
+def register_target_files(
+    repo_path,
+    keystore=None,
+    roles_key_infos=None,
+    commit=False,
+    scheme=DEFAULT_RSA_SIGNATURE_SCHEME,
+    taf_repo=None,
+):
+    """
+    <Purpose>
+        Register all files found in the target directory as targets - updates the targets
+        metadata file, snapshot and timestamp. Sign targets
+        with yubikey if keystore is not provided
+    <Arguments>
+        repo_path:
+        Authentication repository's path
+        keystore:
+        Location of the keystore files
+        roles_key_infos:
+        A dictionary whose keys are role names, while values contain information about the keys.
+        commit_msg:
+        Commit message. If specified, the changes made to the authentication are committed.
+        scheme:
+        A signature scheme used for signing.
+        taf_repo:
+        If taf repository is already initialized, it can be passed and used.
+    """
+    print("Signing target files")
+    roles_key_infos, keystore = _initialize_roles_and_keystore(
+        roles_key_infos, keystore, enter_info=False
+    )
+    roles_infos = roles_key_infos.get("roles")
+    if taf_repo is None:
+        repo_path = Path(repo_path).resolve()
+        taf_repo = Repository(str(repo_path))
+
+    # find files that should be added/modified/removed
+    added_targets_data, removed_targets_data = taf_repo.get_all_target_files_state()
+
+    update_target_metadata(
+        taf_repo,
+        added_targets_data,
+        removed_targets_data,
+        keystore,
+        roles_infos,
+        scheme,
+    )
+
+    if commit:
+        auth_git_repo = GitRepository(path=taf_repo.path)
+        commit_message = input("\nEnter commit message and press ENTER\n\n")
+        auth_git_repo.commit(commit_message)
+
+
+def _setup_role(
+    role_name,
+    threshold,
+    is_yubikey,
+    repository,
+    verification_keys,
+    signing_keys=None,
+    parent=None,
+):
+    role_obj = _role_obj(role_name, repository, parent)
+    role_obj.threshold = threshold
+    if not is_yubikey:
+        for public_key, private_key in zip(verification_keys, signing_keys):
+            role_obj.add_verification_key(public_key)
+            role_obj.load_signing_key(private_key)
+    else:
+        for key_num, key in enumerate(verification_keys):
+            key_name = get_key_name(role_name, key_num, len(verification_keys))
+            role_obj.add_verification_key(key, expires=YUBIKEY_EXPIRATION_DATE)
+            role_obj.add_external_signature_provider(
+                key, partial(yubikey_signature_provider, key_name, key["keyid"])
+            )
+
