+++ taf/api/roles.py
@@ -8,7 +8,12 @@ from taf.hosts import REPOSITORIES_JSON_PATH
 from tuf.repository_tool import TARGETS_DIRECTORY_NAME
 import tuf.roledb
 import taf.repositoriesdb as repositoriesdb
-from taf.keys import get_key_name, load_signing_keys, load_sorted_keys_of_roles
+from taf.keys import (
+    get_key_name,
+    load_signing_keys,
+    load_sorted_keys_of_new_roles,
+    load_sorted_keys_of_roles,
+)
 from taf.api.metadata import update_snapshot_and_timestamp, update_target_metadata
 from taf.auth_repo import AuthenticationRepository
 from taf.constants import (
@@ -93,6 +98,73 @@ def add_role_paths(
         update_snapshot_and_timestamp(auth_repo, keystore, None, None)
         commit_message = input("\nEnter commit message and press ENTER\n\n")
         auth_repo.commit(commit_message)
+
+
+def add_roles(
+    repo_path,
+    keystore=None,
+    roles_key_infos=None,
+    scheme=DEFAULT_RSA_SIGNATURE_SCHEME,
+):
+    yubikeys = defaultdict(dict)
+    auth_repo = AuthenticationRepository(path=repo_path)
+    repo_path = Path(repo_path)
+
+    roles_key_infos, keystore = _initialize_roles_and_keystore(
+        roles_key_infos, keystore
+    )
+
+    new_roles = []
+    taf_repo = Repository(repo_path)
+    existing_roles = taf_repo.get_all_targets_roles()
+    main_roles = ["root", "snapshot", "timestamp", "targets"]
+    existing_roles.extend(main_roles)
+
+    # allow specification of roles without putting them inside targets delegations map
+    # ensuring that it is possible to specify only delegated roles
+    # since creation of delegations expects that structure, place the roles inside targets/delegations
+    delegations_info = {}
+    for role_name, role_data in dict(roles_key_infos["roles"]).items():
+        if role_name not in main_roles:
+            roles_key_infos["roles"].pop(role_name)
+            delegations_info[role_name] = role_data
+    roles_key_infos["roles"].setdefault("targets", {"delegations": {}})[
+        "delegations"
+    ].update(delegations_info)
+
+    # find all existing roles which are parents of the newly added roles
+    # they should be signed after the delegations are created
+    roles = [
+        (role_name, role_data)
+        for role_name, role_data in roles_key_infos["roles"].items()
+    ]
+    parent_roles = set()
+    while len(roles):
+        role_name, role_data = roles.pop()
+        for delegated_role, delegated_role_data in role_data.get(
+            "delegations", {}
+        ).items():
+            if delegated_role not in existing_roles:
+                if role_name not in new_roles:
+                    parent_roles.add(role_name)
+                new_roles.append(delegated_role)
+            roles.append((delegated_role, delegated_role_data))
+
+    if not len(new_roles):
+        print("All roles already set up")
+        return
+
+    repository = taf_repo._repository
+    roles_infos = roles_key_infos.get("roles")
+    signing_keys, verification_keys = load_sorted_keys_of_new_roles(
+        auth_repo, roles_infos, taf_repo, keystore, yubikeys, existing_roles
+    )
+    _create_delegations(
+        roles_infos, repository, verification_keys, signing_keys, existing_roles
+    )
+    for parent_role in parent_roles:
+        _update_role(taf_repo, parent_role, keystore, scheme=scheme)
+    update_snapshot_and_timestamp(taf_repo, keystore, scheme=scheme)
 
 
 def add_signing_key(
